Functions

comododir

Comodo.comododir — Function

comododir()

This function simply returns the string for the Comodo path. This is helpful for instance to load items from the assets` folder.

slidercontrol

Comodo.slidercontrol — Function

slidercontrol(hSlider,ax)

This function adds arrow key control to GLMakie sliders. The inputs are the slider handle hSlider as well as the axis ax. If this function is called the slider can be advanced a step by pressing the right arrow, and returned one step by pressing the left arrow. When one presses and holds the right or left arrow key, the slider will continue to move (as fast as graphics updating is possible on your system) up to the end or start slider position respectively. Users may also use the up or down arrow keys. These function the same as the right and left arrow keys, however, rather than stopping at the slider extrema, the sliders position will "wrap" back to the start when advancing beyond the end position, and vice versa.

elements2indices

Comodo.elements2indices — Function

elements2indices(F)

This function obtains the unique set of indices for the vertices (nodes) used by the the simplices defined by F. The vector F may contain any type of simplices. For instance the elements in F may be of the type GeometryBasics.TriangleFace or GeometryBasics.QuadFace (or any other) for surface mesh data. However volumetric elements of any type are permitted. In essence this function simply returns: julia unique(reduce(vcat,F))Hence any suitable vector containing vectors of numbers permitted byreduce(vcat,F)` is supported.

gridpoints

Comodo.gridpoints — Function

gridpoints(x::Vector{T}, y=x, z=x) where T<:Real

Description: The gridpoints function returns a vector of 3D points which span a grid in 3D space. Points are defined as per the input ranges or range vectors. The output point vector contains elements of the type GeometryBasics.Point3.

Arguments:

interp_biharmonic_spline

Comodo.interp_biharmonic_spline — Function

interp_biharmonic_spline(x,y,xi; extrapolate_method=:linear,pad_data=:linear)

This function uses biharmonic spline interpolation. The input is assumed to represent ordered data representing a curve.

Reference: David T. Sandwell, Biharmonic spline interpolation of GEOS-3 and SEASAT altimeter data, Geophysical Research Letters, 2, 139-142, 1987. doi: 10.1029/GL014i002p00139

interp_biharmonic

Comodo.interp_biharmonic — Function

interp_biharmonic(x,y,xi)

This function uses biharmonic interpolation. The input x should define a vector consisting of m points which are n-dimensional, and the input y should be a vector consisting of m scalar data values.

Reference: David T. Sandwell, Biharmonic spline interpolation of GEOS-3 and SEASAT altimeter data, Geophysical Research Letters, 2, 139-142, 1987. doi: 10.1029/GL014i002p00139

nbezier

Comodo.nbezier — Function

nbezier(P,n)

This function returns n points for an m-th order Bézier spline, based on the m control points contained in the input vector P. This function supports point vectors with elements of the type AbstractPoint{3} (e.g. GeometryBasics.Point{3, Float64}) or Vector{Float64}.

loflinear

Comodo.loftlinear — Function

loftlinear(V1,V2;num_steps=2,close_loop=true,face_type=:tri)

Loft a surface mesh between two input curves

Description

The loftlinear function spans a surface from input curve V1 to curve V2. The surface is formed by "lerping" curves from V1 to V2 in num_loft steps, and forming mesh faces between each curve. If close_loop==true then it is assumed the curves (and therefore the output surface mesh should be closed over, i.e. that a connection should be made between each curve end and start point. The user can request different face types for the output. The default is face_type=:tri which will form isoceles triangles (or equilateral triangles if the spacing is even) for a planar curve. The other face_type options supported are :quad (quadrilateral), and :tri_slash. For the latter, triangles are formed by slashing the quads.

Arguments:

V1::Vector: n-vector

V2::Vector: n-vector

ray_triangle_intersect

Comodo.ray_triangle_intersect — Function

ray_triangle_intersect(f::TriangleFace{Int64},V,ray_origin,ray_vector; rayType = :ray, triSide = 1, tolEps = eps(Float64))

Implementation of the Möller-Trumbore triangle-ray intersection algorithm.

References

Möller, Tomas; Trumbore, Ben (1997). "Fast, Minimum Storage Ray-Triangle Intersection". Journal of Graphics Tools. 2: 21-28. doi:10.1080/10867651.1997.10487468.

mesh_curvature_polynomial

Comodo.mesh_curvature_polynomial — Function

mesh_curvature_polynomial(F,V)

Notes

Implemented with the aid of: 
https://github.com/alecjacobson/geometry-processing-curvature/blob/master/README.md

References

F. Cazals and M. Pouget, "Estimating differential quantities using polynomial 
fitting of osculating jets", Computer Aided Geometric Design, vol. 22, no. 2, 
pp. 121-146, Feb. 2005, doi: 10.1016/j.cagd.2004.09.004.

smoothmesh_hc

Comodo.smoothmesh_hc — Function

smoothmesh_hc(F,V, con_V2V=missing; n=1, α=0.1, β=0.5, tolDist=missing)

Description

This function implements HC (Humphrey's Classes) smoothing. This method uses Laplacian like smoothing but aims to compensate for the shrinkage/swelling seen with pure Laplacian smoothing.

Reference

Vollmer et al. Improved Laplacian Smoothing of Noisy Surface Meshes, 1999 https://doi.org/10.1111/1467-8659.00334

smoothmesh_laplacian

Comodo.smoothmesh_laplacian — Function

smoothmesh_laplacian(F,V,con_V2V=missing; n=1, λ=0.5)

Description

This function implements Weighted Laplacian mesh smoothing. At each iteration, this method replaces each point by the weighted sum of the Laplacian mean for the point and the point itself. The weighting is controlled by the parameter λ which is in the range (0,1). If λ=0 no smoothing occurs. If λ=0 pure Laplacian mean based smoothing occurs. For intermediate values a linear blending between the two occurs.

platonicsolid

Comodo.platonicsolid — Function

platonicsolid(n,r=1.0)

Creates mesh data for a platonic solid of choice

sub2ind

Comodo.sub2ind — Function

sub2ind(siz,A)

Converts the subscript indices in A, for a matrix/array with size siz, to the equivalent linear indices.

ind2sub

Comodo.ind2sub — Function

ind2sub(siz,ind)

Converts the linear indices in `ind`, for a matrix/array with size `siz`, to the equivalent subscript indices.