Functions

loflinear

Comodo.loftlinear — Function

loftlinear(V1,V2;num_steps=2,close_loop=true,face_type=:tri)

Loft a surface mesh between two input curves

Description

The loftlinear function spans a surface from input curve V1 to curve V2. The surface is formed by "lerping" curves form V1 to V2 in num_loft steps, and forming mesh faces between each curve. If close_loop==true then it is assumed the curves and surface should be closed over. The user can request different face types for the output. The default is face_type=:tri which will form isoceles triangles (or equilateral triangles if the spacing is even) for a planar curve. The other face_type options supported are :quad (quadrilateral), and :tri_slash. For the latter, triangles are formed by slashing the quads.

Arguments:

V1::Vector: n-vector

V2::Vector: n-vector

raytriangleintersect

Comodo.ray_triangle_intersect — Function

ray_triangle_intersect(f::TriangleFace{Int64},V,ray_origin,ray_vector; rayType = :ray, triSide = 1, tolEps = eps(Float64))

Implementation of the Möller-Trumbore triangle-ray intersection algorithm.

References

Möller, Tomas; Trumbore, Ben (1997). "Fast, Minimum Storage Ray-Triangle Intersection". Journal of Graphics Tools. 2: 21-28. doi:10.1080/10867651.1997.10487468.

meshcurvaturepolynomial

Comodo.mesh_curvature_polynomial — Function

mesh_curvature_polynomial(F,V)

Notes

Implemented with the aid of: 
https://github.com/alecjacobson/geometry-processing-curvature/blob/master/README.md

References

F. Cazals and M. Pouget, "Estimating differential quantities using polynomial 
fitting of osculating jets", Computer Aided Geometric Design, vol. 22, no. 2, 
pp. 121-146, Feb. 2005, doi: 10.1016/j.cagd.2004.09.004.

smoothmesh_hc

Comodo.smoothmesh_hc — Function

smoothmesh_hc(F,V, con_V2V=missing; n=1, α=0.1, β=0.5, tolDist=missing)

Description

This function implements HC (Humphrey's Classes) smoothing. This method uses
Laplacian like smoothing but aims to compensate for the shrinkage/swelling 
seen with pure Laplacian smoothing.

References

Vollmer et al. Improved Laplacian Smoothing of Noisy Surface Meshes, 1999
https://doi.org/10.1111/1467-8659.00334

smoothmesh_laplacian

Comodo.smoothmesh_laplacian — Function

smoothmesh_laplacian(F,V,con_V2V=missing; n=1, λ=0.5)

Description

This function implements Weighted Laplacian mesh smoothing. At each 
iteration, this method replaces each point by the weighted sum of the 
Laplacian mean for the point and the point itself. The weighting is 
controlled by the parameter λ which is in the range (0,1). If λ=0 no 
smoothing occurs. If λ=0 pure Laplacian mean based smoothing occurs. For 
intermediate values a linear blending between the two occurs.

platonicsolid

Comodo.platonicsolid — Function

platonicsolid(n,r=1.0)

Creates mesh data for a platonic solid of choice

sub2ind

Comodo.sub2ind — Function

sub2ind(siz,A)

Converts the subscript indices in `A`, for a matrix/array with size `siz`, to the equivalent linear indices.

ind2sub

Comodo.ind2sub — Function

ind2sub(siz,ind)

Converts the linear indices in `ind`, for a matrix/array with size `siz`, to the equivalent subscript indices.